+++ R/backwardCompatibility.R
@@ -1,6 +1,27 @@
 #' Backward compabitibility actions:
 #' @export
 backwardCompatibility <- list(
+	renameFunction = list(
+		list(
+			changeVersion = "1.0.23", 
+			functionName = "ICESAcousticCSV", 
+			modelName = "baseline", 
+			newFunctionName = "RstoxData::ICESAcoustic"
+		), 
+		list(
+			changeVersion = "1.0.23", 
+			functionName = "ICESBioticCSV", 
+			modelName = "baseline", 
+			newFunctionName = "RstoxData::ICESBiotic"
+		), 
+		list(
+			changeVersion = "1.0.24", 
+			functionName = "DefineStoxBioticTranslation", 
+			modelName = "baseline", 
+			newFunctionName = "RstoxData::DefineTranslation"
+		)
+	), 
+	
 	removeParameter = list(
 		list(
 			changeVersion = "1.0.18", 
@@ -43,21 +64,57 @@ backwardCompatibility <- list(
 			functionName = "ICESDatras", 
 			modelName = "baseline", 
 			parameterName = "AddStationType"
+		), 
+		list(
+			changeVersion = "1.0.24", 
+			functionName = "TranslateStoxBiotic", 
+			modelName = "baseline", 
+			parameterName = "Translation"
+		), 
+		list(
+			changeVersion = "1.0.24", 
+			functionName = "TranslateStoxBiotic", 
+			modelName = "baseline", 
+			parameterName = "TranslationDefinition"
 		)
 	),  
 	
-	renameFunction = list(
+	renameParameter = list(
 		list(
-			changeVersion = "1.0.23", 
-			functionName = "ICESAcousticCSV", 
+			changeVersion = "1.0.24", 
+			functionName = "TranslateStoxBiotic", 
 			modelName = "baseline", 
-			newFunctionName = "RstoxData::ICESAcoustic"
+			parameterName = "StoxBioticTranslation",
+			newParameterName = "Translation"
 		), 
 		list(
-			changeVersion = "1.0.23", 
-			functionName = "ICESBioticCSV", 
+			changeVersion = "1.0.24", 
+			functionName = "DefineTranslation", 
 			modelName = "baseline", 
-			newFunctionName = "RstoxData::ICESBiotic"
+			parameterName = "Translation",
+			newParameterName = "TranslationTable"
+		)
+	),  
+	
+	translateParameter = list(
+		list(
+			changeVersion = "1.0.24", 
+			functionName = "DefineTranslation", 
+			modelName = "baseline", 
+			parameterName = "DefinitionMethod", 
+			value = "Table", 
+			newValue = "TranslationTable"
+		)
+	), 
+	
+	renameProcessData = list(
+		list(
+			changeVersion = "1.0.24", 
+			functionName = "DefineTranslation", 
+			modelName = "baseline", 
+			processDataName = "StoxBioticTranslation",
+			newProcessDataName = "Translation"
 		)
 	)
+	
 )